add documentation and refactor
