+++ src/zrb/builtin/project/add/fastapp/app/template/src/kebab-zrb-app-name/src/main.py
@@ -1,9 +1,7 @@
-from integration.app import app
+from integration.app.app import app
 from module.auth.register_module import register_auth
 from module.log.register_module import register_log
 
-# Make sure app is loaded.
-# Uvicorn or adny ASGII server you use will pick it up and run the app.
 assert app
 register_auth()
 register_log()
